Fixed compile errors cryptonate.sol
This is a pretty big change to some things so we can roll it back if we don't like it.
I put the whole code inside a contract, because that is how Solidity works in my understanding. I named the contract Cryptonate, as well as changing the name of this file to Cryptonate.sol
I also ran the code through Remix, the online Solidity IDE and fixed all the compile errors that came up. There were some small typos and misnamed variables, so I fixed those, giving preference to the variable name that made the most sense for what it was being used.